backend/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-cors==3.0.10
pyodbc==4.0.32
python-dotenv==0.19.0
flask-jwt-extended==4.3.1
Werkzeug==2.0.1